+++ ft_strmapi.c
@@ -6,7 +6,7 @@
 /*   By: lilefebv <lilefebv@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/08 10:48:30 by lilefebv          #+#    #+#             */
-/*   Updated: 2024/11/08 12:29:51 by lilefebv         ###   ########lyon.fr   */
+/*   Updated: 2024/11/10 15:13:36 by lilefebv         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
-	unsigned int	i;
-	char			*new_str;
+	size_t	i;
+	char	*new_str;
 
 	new_str = malloc(sizeof(char) * (ft_strlen(s) + 1));
 	if (!new_str)
